Move check/whitespace_widths to `Universal` profile
Also added rationale text.

com.google.fonts/check/whitespace_widths moved from `OpenType` profile.
(issue #3843)
