+++ Lib/fontbakery/profiles/opentype.py
@@ -13,7 +13,6 @@ OPENTYPE_PROFILE_IMPORTS = (
         "loca",
         "hhea",
         "dsig",
-        "hmtx",
         "gdef",
         "gpos",
         "kern",
@@ -62,7 +61,6 @@ OPENTYPE_PROFILE_CHECKS = [
     'com.google.fonts/check/linegaps',
     'com.google.fonts/check/unitsperem',
     'com.google.fonts/check/dsig',
-    'com.google.fonts/check/whitespace_widths',
     'com.google.fonts/check/gdef_spacing_marks',
     'com.google.fonts/check/gdef_mark_chars',
     'com.google.fonts/check/gdef_non_mark_chars',
